Moved function into correct folder
